services/nlp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxsentinel-nlp"
version = "0.1.0"
description = "NLP, Keyword, Sentiment, and PII service for VoxSentinel"
requires-python = ">=3.12"
dependencies = [
    "tg-common",
    "pyahocorasick>=2.1",
    "rapidfuzz>=3.8",
    "transformers>=4.40",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "gliner>=0.2",
    "torch>=2.3",
    "structlog>=24.2",
    "prometheus-client>=0.20",
    "redis>=5.0",
    "uvicorn>=0.30",
    "fastapi>=0.111",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
